% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_snps.R
\name{clump_gxe_snps}
\alias{clump_gxe_snps}
\title{Clump significant GxE SNPs}
\usage{
clump_gxe_snps(
  gxe_res,
  p_cutoff = 5e-08,
  clump_kb = 10000,
  clump_r2 = 0.001,
  pop = "EUR",
  bfile = NULL,
  plink_bin = "plink"
)
}
\arguments{
\item{gxe_res}{fastGxE GxE result file or data.frame.}

\item{p_cutoff}{P-value threshold for GxE.}

\item{clump_kb}{Clumping window.}

\item{clump_r2}{LD r2 threshold.}

\item{pop}{Population for TwoSampleMR clumping.}

\item{bfile}{Reference PLINK bfile for clumping.}

\item{plink_bin}{PLINK binary.}
}
\value{
A clumped data.frame.
}
\description{
Clump significant GxE SNPs
}
