% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{compute_main_prs}
\alias{compute_main_prs}
\title{Compute main-effect PRS from genotype dosage and SNP weights}
\usage{
compute_main_prs(
  geno,
  weights,
  id_col = "ID",
  snp_col = "SNP",
  beta_col = "beta"
)
}
\arguments{
\item{geno}{Genotype data.frame with ID and SNP dosage columns.}

\item{weights}{data.frame with SNP and beta columns.}

\item{id_col}{ID column.}

\item{snp_col}{SNP column in weights.}

\item{beta_col}{beta column in weights.}
}
\value{
data.frame with ID and PRS.
}
\description{
Compute main-effect PRS from genotype dosage and SNP weights
}
