% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmr.R
\name{create_nlmr_summary}
\alias{create_nlmr_summary}
\title{Create nonlinear MR summary}
\usage{
create_nlmr_summary(
  data,
  outcome_col = "outcome",
  exposure_col = "exposure_adjusted",
  prs_col = "PRS",
  covar_cols = NULL,
  family = c("gaussian", "binomial"),
  strata_method = c("ranked", "residual"),
  q = 10
)
}
\arguments{
\item{data}{Analysis data.frame.}

\item{outcome_col}{Outcome column.}

\item{exposure_col}{Exposure column.}

\item{prs_col}{PRS column.}

\item{covar_cols}{Covariate columns.}

\item{family}{"gaussian" or "binomial".}

\item{strata_method}{"ranked" or "residual".}

\item{q}{Number of strata.}
}
\value{
Output from SUMnlmr::create_nlmr_summary.
}
\description{
Create nonlinear MR summary
}
