% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_extract.R
\name{extract_genotypes_plink}
\alias{extract_genotypes_plink}
\title{Extract SNP genotypes using PLINK and recode to additive dosage}
\usage{
extract_genotypes_plink(
  bfile_prefixes,
  snps,
  keep_file = NULL,
  out_prefix,
  plink_bin = "plink",
  verbose = TRUE
)
}
\arguments{
\item{bfile_prefixes}{Character vector of PLINK bfile prefixes, e.g. chr1 to chr22.}

\item{snps}{Character vector of SNP IDs.}

\item{keep_file}{Optional PLINK keep file.}

\item{out_prefix}{Output prefix.}

\item{plink_bin}{PLINK executable.}

\item{verbose}{Print command.}
}
\value{
Path to PLINK raw file.
}
\description{
Extract SNP genotypes using PLINK and recode to additive dosage
}
