% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gxe_expand_col_range}
\alias{gxe_expand_col_range}
\title{Expand phenotype column range such as PC1:PC20 or Age:sleep}
\usage{
gxe_expand_col_range(x, col_names)
}
\description{
Expand phenotype column range such as PC1:PC20 or Age:sleep
}
\keyword{internal}
