% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_exposure.R
\name{make_adjusted_exposure}
\alias{make_adjusted_exposure}
\title{Make GxE-adjusted exposure}
\usage{
make_adjusted_exposure(
  exposure,
  adjustment,
  id_col = "ID",
  exposure_col = "exposure",
  adjustment_col = "GxE_adjustment"
)
}
\arguments{
\item{exposure}{data.frame with ID and exposure.}

\item{adjustment}{data.frame from gxe_compute_gxe_adjustment.}

\item{id_col}{ID column.}

\item{exposure_col}{Exposure column.}

\item{adjustment_col}{Adjustment column.}
}
\value{
data.frame with ID, exposure, GxE_adjustment, exposure_adjusted.
}
\description{
Computes:
\deqn{
  X' = X - A
}
}
