% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heterogeneity}
\alias{plot_heterogeneity}
\title{Plot genetic association heterogeneity}
\usage{
plot_heterogeneity(heter_data, exposure = "Exposure")
}
\arguments{
\item{heter_data}{data.frame with beta, lci, uci, strata, Method.}

\item{exposure}{Exposure name.}
}
\value{
ggplot object.
}
\description{
Plot genetic association heterogeneity
}
