% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lace}
\alias{plot_lace}
\title{Plot LACE estimates across strata}
\usage{
plot_lace(
  lace_data,
  exposure = "Exposure",
  outcome = "Outcome",
  family = c("gaussian", "binomial")
)
}
\arguments{
\item{lace_data}{data.frame with beta, lci, uci, strata, Method.}

\item{exposure}{Exposure name.}

\item{outcome}{Outcome name.}

\item{family}{"gaussian" or "binomial".}
}
\value{
ggplot object.
}
\description{
Plot LACE estimates across strata
}
