% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastgxe.R
\name{run_fastgxe}
\alias{run_fastgxe}
\title{Run fastGxE and optional mmSuSiE}
\usage{
run_fastgxe(
  bfile,
  pheno,
  trait,
  env_int,
  covar = NULL,
  class_covar = NULL,
  outdir = "output",
  prefix = "test",
  fastgxe_path = "fastgxe",
  python_path = "python",
  grm_prefix_existing = NULL,
  run_grm_reformat = TRUE,
  grm_npart = NULL,
  threads = NULL,
  cut_value = 0.05,
  no_standardize_env = FALSE,
  no_noisebye = FALSE,
  run_mmsusie = TRUE,
  mmsusie_snp = NULL,
  mmsusie_p_cutoff = 5e-08,
  mmsusie_ld_r2 = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{bfile}{PLINK bfile prefix or .bed file.}

\item{pheno}{Phenotype file path or data.frame.}

\item{trait}{Trait column name.}

\item{env_int}{Environmental variables used for GxE interaction. Supports ranges such as Age:sleep.}

\item{covar}{Covariates, e.g. "PC1:PC20".}

\item{class_covar}{Categorical covariates.}

\item{outdir}{Output directory.}

\item{prefix}{Output prefix.}

\item{fastgxe_path}{Path to fastGxE executable.}

\item{python_path}{Path to Python.}

\item{grm_prefix_existing}{Existing GRM prefix. If NULL, GRM will be built.}

\item{run_grm_reformat}{Whether to reformat GRM for mmSuSiE.}

\item{grm_npart}{Number of GRM partitions.}

\item{threads}{Number of threads.}

\item{cut_value}{Cut value for GRM grouping.}

\item{no_standardize_env}{Whether to disable environment standardization.}

\item{no_noisebye}{Whether to disable NoiseBye.}

\item{run_mmsusie}{Whether to run mmSuSiE.}

\item{mmsusie_snp}{User-specified SNP for mmSuSiE.}

\item{mmsusie_p_cutoff}{P-value cutoff for mmSuSiE lead SNP extraction.}

\item{mmsusie_ld_r2}{LD r2 threshold.}

\item{verbose}{Print commands.}
}
\value{
A list of output file paths.
}
\description{
This function wraps the fastGxE workflow:
make GRM, process GRM, run GxE scan, optionally reformat GRM and run mmSuSiE.
}
