% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_snps.R
\name{select_main_snps}
\alias{select_main_snps}
\title{Select genome-wide significant main-effect SNPs}
\usage{
select_main_snps(
  main_res,
  p_cutoff = 5e-08,
  snp_col = "SNP",
  beta_col = "beta",
  p_col = "p"
)
}
\arguments{
\item{main_res}{fastGxE main-effect result file or data.frame.}

\item{p_cutoff}{P-value threshold.}

\item{snp_col}{SNP column name.}

\item{beta_col}{Beta column name.}

\item{p_col}{P-value column name.}
}
\value{
A data.frame of significant SNPs.
}
\description{
Select genome-wide significant main-effect SNPs
}
